backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
pymongo==4.6.1
python-dotenv==1.0.0
pytesseract==0.3.10
Pillow==10.1.0
spacy==3.7.2
scikit-learn==1.3.2